pyproject.toml
[project]
name = "phspy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphql-core==3.2.3",
    "pytest==8.2.1",
    "requests==2.32.3",
    "requests-mock==1.12.1",
]
